LUCENE-1257: More Java5 updates
git-svn-id: https://svn.apache.org/repos/asf/lucene/java/trunk@829013 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/lucene/index/BufferedDeletes.java
@@ -136,7 +136,7 @@ class BufferedDeletes {
     if (queries.size() > 0) {
       newDeleteQueries = new HashMap<Query, Integer>(queries.size());
       for(Entry<Query,Integer> entry: queries.entrySet()) {
-        Integer num = (Integer) entry.getValue();
+        Integer num = entry.getValue();
         newDeleteQueries.put(entry.getKey(),
                              Integer.valueOf(mapper.remap(num.intValue())));
       }
